+++ packages/framework/src/evaluate/binding.ts
@@ -82,9 +82,10 @@ export const createBindingAtom = <T = unknown>(
     const evaluationContext = createEvaluationContext({
       applicationContext: {
         application: {
-          theme: get(themeAtom),
           languages: appData.application?.languages,
+          themes: appData.application?.themes,
         },
+        selectedTheme: get(themeAtom),
         env: get(envAtom),
         secrets: get(secretAtom),
       },