+++ packages/framework/src/evaluate/context.ts
@@ -1,19 +1,23 @@
-import { mapKeys, merge } from "lodash-es";
+import { mapKeys, merge, keys } from "lodash-es";
 import type { EnsembleInterface } from "../shared/ensemble";
 import type {
   ApplicationContextDefinition,
   ScreenContextDefinition,
 } from "../state";
-import type { EnsembleAppModel } from "../shared";
+import type { EnsembleAppModel, EnsembleThemeModel } from "../shared";
 
 export interface EvaluationContextProps {
   applicationContext: Omit<
     Partial<ApplicationContextDefinition>,
     "application"
-  > & { application?: Partial<EnsembleAppModel> | null };
+  > & {
+    application?: Partial<EnsembleAppModel> | null;
+    selectedTheme?: EnsembleThemeModel;
+  };
   screenContext: Partial<ScreenContextDefinition>;
   ensemble: Partial<EnsembleInterface>;
   context?: { [key: string]: unknown };
+  app?: { [key: string]: unknown };
 }
 
 export const createEvaluationContext = ({
@@ -22,7 +26,7 @@ export const createEvaluationContext = ({
   ensemble,
   context,
 }: EvaluationContextProps): { [key: string]: unknown } => {
-  const theme = applicationContext.application?.theme;
+  const theme = applicationContext.selectedTheme;
   const appInputs = merge(
     {},
     applicationContext.env,
@@ -40,7 +44,10 @@ export const createEvaluationContext = ({
 
   const app = {
     languages: applicationContext.application?.languages,
+    themes: keys(applicationContext.application?.themes),
+    theme: applicationContext.selectedTheme?.name,
   };
   const env = applicationContext.env;
+
   return merge({}, { app, ensemble, env }, appInputs, screenInputs, context);
 };