+++ packages/framework/src/state/application.ts
@@ -1,8 +1,9 @@
-import { atomWithStorage, createJSONStorage } from "jotai/utils";
 import { atom } from "jotai";
 import { focusAtom } from "jotai-optics";
 import type { EnsembleAppModel, EnsembleThemeModel } from "../shared";
 import type { EnsembleUser } from "./user";
+
+export const defaultThemeDefinition = { name: "default" };
 
 export interface ApplicationContextDefinition {
   application: EnsembleAppModel | null;
@@ -11,6 +12,7 @@ export interface ApplicationContextDefinition {
   env: { [key: string]: unknown };
   auth: unknown;
   user: EnsembleUser | null;
+  useMockResponse: boolean;
 }
 
 export interface ApplicationContextActions {
@@ -24,20 +26,14 @@ export const defaultApplicationContext = {
   auth: null,
   user: null,
   secrets: {},
+  useMockResponse: false,
 };
 
 export const appAtom = atom<ApplicationContextDefinition>(
   defaultApplicationContext,
 );
 
-export const themeAtom = atom<EnsembleThemeModel | undefined>(undefined);
-
-const backingStorage = createJSONStorage<string>(() => sessionStorage);
-export const selectedThemeNameAtom = atomWithStorage<string>(
-  "ensemble.selectedThemeName",
-  "default",
-  backingStorage,
-);
+export const themeAtom = atom<EnsembleThemeModel>(defaultThemeDefinition);
 
 export const envAtom = focusAtom(appAtom, (optic) => optic.prop("env"));