+++ packages/framework/src/state/screen.ts
@@ -1,26 +1,20 @@
-import { atom, useAtom, useAtomValue } from "jotai";
+import { atom } from "jotai";
 import { focusAtom } from "jotai-optics";
-import { clone } from "lodash-es";
-import { useCallback } from "react";
-import isEqual from "react-fast-compare";
 import type { Response, WebSocketConnection } from "../data";
-import type {
-  EnsembleAPIModel,
-  EnsembleAppModel,
-  EnsembleScreenModel,
-  EnsembleSocketModel,
-} from "../shared";
+import type { EnsembleAppModel, EnsembleScreenModel } from "../shared";
 import type { WidgetState } from "./widget";
 
 export interface ScreenContextDefinition {
   app?: EnsembleAppModel;
   model?: EnsembleScreenModel;
-  data: {
-    [key: string]: Response | WebSocketConnection;
-  };
+  data: ScreenContextData;
   widgets: { [key: string]: WidgetState };
   inputs?: { [key: string]: unknown };
   [key: string]: unknown;
+}
+
+export interface ScreenContextData {
+  [key: string]: Response | WebSocketConnection;
 }
 
 export interface ScreenContextActions {
@@ -60,30 +54,3 @@ export const screenGlobalScriptAtom = focusAtom(screenAtom, (optic) =>
 export const screenImportScriptAtom = focusAtom(screenAtom, (optic) =>
   optic.prop("model").optional().prop("importedScripts"),
 );
-
-export const useScreenData = (): { apis?: EnsembleAPIModel[] } & {
-  sockets?: EnsembleSocketModel[];
-} & Pick<ScreenContextDefinition, "data"> &
-  Pick<ScreenContextActions, "setData"> => {
-  const apis = useAtomValue(screenApiAtom);
-  const sockets = useAtomValue(screenSocketAtom);
-  const [data, setDataAtom] = useAtom(screenDataAtom);
-
-  const setData = useCallback(
-    (name: string, response: Response | WebSocketConnection) => {
-      if (isEqual(data[name], response)) {
-        return;
-      }
-      data[name] = response;
-      setDataAtom(clone(data));
-    },
-    [data, setDataAtom],
-  );
-
-  return {
-    apis,
-    sockets,
-    data,
-    setData,
-  };
-};