+++ packages/runtime/src/widgets/DataGrid/DataGrid.tsx
@@ -137,6 +137,10 @@ const ResizableTitle: React.FC<ResizableProps & ResizableState> = (props) => {
   );
 };
 
+const defaultGridMutatorOptions = {
+  suppressCallbacks: false,
+};
+
 export const DataGrid: React.FC<GridProps> = (props) => {
   const {
     "item-template": itemTemplate,
@@ -251,18 +255,22 @@ export const DataGrid: React.FC<GridProps> = (props) => {
 
   // update page number callback
   const updatePageNumber = useCallback(
-    (newPage: number) => {
+    (newPage: number, options = defaultGridMutatorOptions) => {
       setCurPage(newPage);
-      onPageChangeActionCallback(newPage, pageSize || 10);
+      if (!options.suppressCallbacks) {
+        onPageChangeActionCallback(newPage, pageSize || 10);
+      }
     },
     [pageSize, onPageChangeActionCallback],
   );
 
   // update pageSize callback
   const updatePageSize = useCallback(
-    (newPageSize: number) => {
+    (newPageSize: number, options = defaultGridMutatorOptions) => {
       setPageSize(newPageSize);
-      onPageChangeActionCallback(curPage, newPageSize || 10);
+      if (!options.suppressCallbacks) {
+        onPageChangeActionCallback(curPage, newPageSize || 10);
+      }
     },
     [curPage, onPageChangeActionCallback],
   );