+++ packages/runtime/src/widgets/Form/Date/Date.tsx
@@ -1,8 +1,16 @@
-import { useRegisterBindings } from "@ensembleui/react-framework";
+import { error, useRegisterBindings } from "@ensembleui/react-framework";
 import type { Expression, EnsembleAction } from "@ensembleui/react-framework";
 import { DatePicker, Form } from "antd";
 import { useState, useCallback, useEffect } from "react";
+import type { Dayjs } from "dayjs";
 import dayjs from "dayjs";
+import advancedFormat from "dayjs/plugin/advancedFormat";
+import customParseFormat from "dayjs/plugin/customParseFormat";
+import localeData from "dayjs/plugin/localeData";
+import weekday from "dayjs/plugin/weekday";
+import weekOfYear from "dayjs/plugin/weekOfYear";
+import weekYear from "dayjs/plugin/weekYear";
+import { isArray } from "lodash-es";
 import { WidgetRegistry } from "../../../registry";
 import type { EnsembleWidgetProps } from "../../../shared/types";
 import { useEnsembleAction } from "../../../runtime/hooks/useEnsembleAction";
@@ -10,9 +18,18 @@ import type { FormInputProps } from "../types";
 import { EnsembleFormItem } from "../FormItem";
 import { DateDisplayFormat } from "./utils/DateConstants";
 
+/* eslint-disable import/no-named-as-default-member */
+dayjs.extend(customParseFormat);
+dayjs.extend(advancedFormat);
+dayjs.extend(weekday);
+dayjs.extend(localeData);
+dayjs.extend(weekOfYear);
+dayjs.extend(weekYear);
+/* eslint-enable import/no-named-as-default-member */
+
 const widgetName = "Date";
 
-const standardizeTimestamp = (timestamp: string, format?: string) =>
+const standardizeTimestamp = (timestamp: string, format?: string): string =>
   dayjs(timestamp).format(format || DateDisplayFormat);
 
 export type DateProps = {
@@ -29,7 +46,7 @@ export const Date: React.FC<DateProps> = (props) => {
   const [value, setValue] = useState<string>();
   const action = useEnsembleAction(props.onChange);
 
-  const { id, values } = useRegisterBindings(
+  const { id, values, rootRef } = useRegisterBindings(
     {
       ...props,
       initialValue: props.initialValue ?? props.value,
@@ -46,7 +63,7 @@ export const Date: React.FC<DateProps> = (props) => {
     if (values?.initialValue) {
       setValue(standardizeTimestamp(values.initialValue, values.format));
     }
-  }, [values?.initialValue]);
+  }, [values?.format, values?.initialValue]);
 
   const onChangeCallback = useCallback(
     (date?: string) => {
@@ -64,9 +81,13 @@ export const Date: React.FC<DateProps> = (props) => {
     [action, id, props],
   );
 
-  const onDateChange = (date: string): void => {
-    setValue(date);
-    const formattedDate = standardizeTimestamp(date, values?.format);
+  const onDateChange = (date: Dayjs, dateString: string | string[]): void => {
+    if (isArray(dateString)) {
+      error("Received an array of dates when only expecting one");
+      return;
+    }
+    setValue(dateString);
+    const formattedDate = standardizeTimestamp(dateString, values?.format);
     if (formattedDate === "Invalid Date") {
       onChangeCallback("");
     } else {
@@ -82,13 +103,11 @@ export const Date: React.FC<DateProps> = (props) => {
         [values?.id ?? values?.label ?? ""]: value ? dayjs(value) : undefined,
       });
     }
-  }, [value, formInstance]);
+  }, [value, formInstance, values?.id, values?.label]);
 
   return (
     <EnsembleFormItem
-      initialValue={
-        values?.initialValue ? dayjs(values.initialValue) : undefined
-      }
+      initialValue={values?.initialValue ? dayjs(values.initialValue) : null}
       values={values}
     >
       <DatePicker
@@ -100,8 +119,9 @@ export const Date: React.FC<DateProps> = (props) => {
         format={values?.format}
         onChange={onDateChange}
         placeholder={values?.hintText}
+        ref={rootRef}
         style={{ width: "100%", ...values?.styles }}
-        value={String(dayjs(values?.value))}
+        value={values?.value ? dayjs(values.value) : null}
         {...(values?.showCalendarIcon === false
           ? { suffixIcon: false }
           : undefined)}